break down program into multiple functions
C:/Users/kushal/.cargo/bin/cargo.exe run --color=always --package helloworld --bin helloworld
    Finished `dev` profile [unoptimized + debuginfo] target(s) in 5.38s
     Running `target\debug\helloworld.exe`
Guess the number!
Please input your guess.
1
You guessed: 1
Too small!
The sum of your guess and the secret number is: 98
The product of your guess and the secret number is: 97
The greatest common divisor of your guess and the secret number is: 1
Please input your guess.
99
You guessed: 99
Too big!
The sum of your guess and the secret number is: 196
The product of your guess and the secret number is: 9603
The greatest common divisor of your guess and the secret number is: 1
Please input your guess.
98
You guessed: 98
Too big!
The sum of your guess and the secret number is: 195
The product of your guess and the secret number is: 9506
The greatest common divisor of your guess and the secret number is: 1
Please input your guess.
97
You guessed: 97
You win!
Press Enter to exit...

there is a defect though. I don't see "hello world" anywhere here